requirements.txt
# Environment Information
# OS: Linux (Arch)
# Python: 3.12.12
# CUDA: 13.1
# GPU: NVIDIA GeForce RTX 5060 Laptop GPU
# TensorFlow: 2.21.0-dev20260109
# PyTorch: Not installed (add torch if needed)

tensorflow
numpy
opencv-python
PyYAML
